fix: gateway has to download the test data
