+++ crates/rattler_repodata_gateway/src/utils/mod.rs
@@ -1,8 +1,9 @@
+use std::fmt::Write;
+
 use ::url::Url;
 pub use body::BodyStreamExt;
 pub use encoding::{AsyncEncoding, Encoding};
 pub use flock::LockedFile;
-use std::fmt::Write;
 
 mod encoding;
 
@@ -15,11 +16,11 @@ mod flock;
 /// Convert a URL to a cache filename
 pub(crate) fn url_to_cache_filename(url: &Url) -> String {
     // Start Rant:
-    // This function mimics behavior from Mamba which itself mimics this behavior from Conda.
-    // However, I find this function absolutely ridiculous, it contains all sort of weird edge
-    // cases and returns a hash that could very easily collide with other files. And why? Why not
-    // simply return a little more descriptive? Like a URL encoded string?
-    // End Rant.
+    // This function mimics behavior from Mamba which itself mimics this behavior
+    // from Conda. However, I find this function absolutely ridiculous, it
+    // contains all sort of weird edge cases and returns a hash that could very
+    // easily collide with other files. And why? Why not simply return a little
+    // more descriptive? Like a URL encoded string? End Rant.
     let mut url_str = url.to_string();
 
     // Ensure there is a slash if the URL is empty or doesnt refer to json file
@@ -42,9 +43,12 @@ pub(crate) fn url_to_cache_filename(url: &Url) -> String {
 }
 
 #[cfg(test)]
-mod test {
-    use super::url_to_cache_filename;
+pub(crate) mod test {
+    use std::path::{Path, PathBuf};
+
     use url::Url;
+
+    use super::url_to_cache_filename;
 
     #[test]
     fn test_url_to_cache_filename() {
@@ -52,5 +56,27 @@ mod test {
             url_to_cache_filename(&Url::parse("http://test.com/1234/").unwrap()),
             "302f0a61"
         );
+    }
+
+    pub(crate) fn test_dir() -> PathBuf {
+        Path::new(env!("CARGO_MANIFEST_DIR")).join("../../test-data")
+    }
+
+    pub(crate) async fn fetch_repo_data(subdir: &str) -> Result<(), reqwest::Error> {
+        let path = test_dir().join(format!("channels/conda-forge/{}/repodata.json", subdir));
+        if path.exists() {
+            return Ok(());
+        } else {
+            tokio::fs::create_dir_all(path.parent().unwrap())
+                .await
+                .unwrap();
+        }
+        let data = reqwest::get(format!(
+            "https://rattler-test.pixi.run/test-data/channels/conda-forge/{}/repodata.json",
+            subdir
+        ))
+        .await?;
+        tokio::fs::write(path, data.bytes().await?).await.unwrap();
+        Ok(())
     }
 }
